更新i18n和操作指引
